embed cache on core
