--- packages/cache/package.json
@@ -1,27 +1,0 @@
-{
-  "name": "@grafoo/cache",
-  "version": "0.0.1",
-  "description": "the grafoo client cache",
-  "main": "dist/index.js",
-  "typings": "dist/index.d.ts",
-  "author": "malbernaz<albernazmiguel@gmail.com>",
-  "license": "MIT",
-  "scripts": {
-    "prepare": "grafoo-bundle --input src/index.ts --skip-compression",
-    "test": "jest",
-    "test:coverage": "jest --coverage && codecov"
-  },
-  "jest": {
-    "moduleFileExtensions": [
-      "ts",
-      "tsx",
-      "js"
-    ],
-    "testMatch": [
-      "**/__tests__/*.+(ts|tsx|js)"
-    ],
-    "transform": {
-      "^.+\\.(ts|tsx|js)$": "<rootDir>/../../scripts/jest-setup.js"
-    }
-  }
-}
