+++ packages/core/__tests__/map-objects.ts
