+++ packages/core/package.json
@@ -7,6 +7,22 @@
   "author": "malbernaz<albernazmiguel@gmail.com>",
   "license": "MIT",
   "scripts": {
-    "prepare": "grafoo-bundle --input src/index.ts"
+    "prepare": "grafoo-bundle --input src/index.ts",
+    "test": "jest",
+    "test:coverage": "jest --coverage && codecov"
+  },
+  "jest": {
+    "moduleFileExtensions": [
+      "ts",
+      "tsx",
+      "js",
+      "json"
+    ],
+    "testMatch": [
+      "**/__tests__/*.+(ts|tsx|js)"
+    ],
+    "transform": {
+      "^.+\\.(ts|tsx|js)$": "<rootDir>/../../scripts/jest-setup.js"
+    }
   }
 }
