+++ packages/core/src/map-objects.ts
