+++ packages/types/types/index.ts
@@ -1,5 +1,4 @@
 export * from "./bindings";
-export * from "./cache";
 export * from "./core";
 export * from "./react";
 export * from "./tag";
